Refactor backend entities and services to use Long IDs instead of Integer, update QuizResult processing, and remove unused DTOs
- Changed ID types from Integer to Long in PracticeSession, PracticeSessionFlashCard, and related entities.
- Updated controller methods to accept Long IDs for session and student identifiers.
- Added new methods in PracticeSessionService for starting and ending sessions, and checking ongoing sessions.
- Implemented quiz submission processing in QuizResultService, including score calculation and session association.
- Removed QuizSubmissionDTO and its inner class as they are no longer needed.
- Updated FlashCardService to handle Long IDs and added methods for filtering flashcards by difficulty.
- Adjusted enums for difficulty levels and ratings to better reflect their purpose.
- Cleaned up unnecessary imports and added JSON ignore properties to prevent serialization issues.

+++ BACKEND/src/main/java/com/guilherme/project/languageplatform/controller/PracticeSessionController.java
@@ -22,12 +22,12 @@ public class PracticeSessionController {
     }
 
     @GetMapping("/{id}")
-    public Optional<PracticeSession> getSessionById(@PathVariable Integer id) {
+    public Optional<PracticeSession> getSessionById(@PathVariable Long id) {
         return practiceSessionService.getPracticeSessionById(id);
     }
 
     @GetMapping("/ongoing/{studentId}")
-    public ResponseEntity<PracticeSession> getOngoingSession(@PathVariable Integer studentId) {
+    public ResponseEntity<PracticeSession> getOngoingSession(@PathVariable Long studentId) {
         return practiceSessionService.getOngoingSessionByStudentId(studentId)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -39,12 +39,12 @@ public class PracticeSessionController {
     }
 
     @PutMapping("/{id}")
-    public PracticeSession updateSession(@PathVariable Integer id, @RequestBody PracticeSession updatedSession) {
+    public PracticeSession updateSession(@PathVariable Long id, @RequestBody PracticeSession updatedSession) {
         return practiceSessionService.updatePracticeSession(id, updatedSession);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteSession(@PathVariable Integer id) {
+    public void deleteSession(@PathVariable Long id) {
         practiceSessionService.deletePracticeSession(id);
     }
 }
